+++ src/main/java/li/cil/sedna/api/memory/MemoryMap.java
@@ -5,7 +5,7 @@ import li.cil.sedna.api.device.MemoryMappedDevice;
 
 import javax.annotation.Nullable;
 import java.util.Optional;
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 /**
  * Represents a physical memory mapping of devices.
@@ -19,21 +19,21 @@ public interface MemoryMap {
      * @param size  the size of the memory range to find.
      * @return the address of a free memory range, if one was found.
      */
-    OptionalInt findFreeRange(final int start, final int end, final int size);
+    OptionalLong findFreeRange(final long start, final long end, final int size);
 
     /**
      * Tries to add a new device to the mapping at the specified address.
      * <p>
      * This can fail if the new device would overlap a memory range already occupied by
      * an existing device. In that case this method will return <code>false</code>. Use
-     * {@link #findFreeRange(int, int, int)} to obtain an address the device can be
+     * {@link #findFreeRange(long, long, int)} to obtain an address the device can be
      * added at.
      *
      * @param address the address to add the device add.
      * @param device  the device to add.
      * @return <code>true</code> if the device was added; <code>false</code> otherwise.
      */
-    boolean addDevice(final int address, final MemoryMappedDevice device);
+    boolean addDevice(final long address, final MemoryMappedDevice device);
 
     /**
      * Removes a device from the memory map.
@@ -57,14 +57,14 @@ public interface MemoryMap {
      * <p>
      * This is useful for getting a direct reference to a {@link MemoryMappedDevice} at
      * a specific memory location, usually to perform multiple read or write operations
-     * on it without having to go through the slower {@link #load(int, int)} and
-     * {@link #store(int, long, int)} calls.
+     * on it without having to go through the slower {@link #load(long, int)} and
+     * {@link #store(long, long, int)} calls.
      *
      * @param address the address to get a memory range for.
      * @return the memory range the address falls into, if any.
      */
     @Nullable
-    MemoryRange getMemoryRange(final int address);
+    MemoryRange getMemoryRange(final long address);
 
     /**
      * Marks a location in memory dirty.
@@ -82,26 +82,26 @@ public interface MemoryMap {
      * <p>
      * When performing many operations on an address range that is known to be occupied by
      * a single device, it is more efficient to obtain a reference to that device via
-     * {@link #getMemoryRange(int)} instead and operate on that device directly.
+     * {@link #getMemoryRange(long)} instead and operate on that device directly.
      *
      * @param address  the physical address to read a value from.
      * @param sizeLog2 the size of the value to read. See {@link Sizes}.
      * @return the value read from the specified location.
      * @throws MemoryAccessException if an error occurred accessing the memory a the specified location.
      */
-    long load(final int address, final int sizeLog2) throws MemoryAccessException;
+    long load(final long address, final int sizeLog2) throws MemoryAccessException;
 
     /**
      * Writes a value to the specified physical address.
      * <p>
      * When performing many operations on an address range that is known to be occupied by
      * a single device, it is more efficient to obtain a reference to that device via
-     * {@link #getMemoryRange(int)} instead and operate on that device directly.
+     * {@link #getMemoryRange(long)} instead and operate on that device directly.
      *
      * @param address  the physical address to write a value to.
      * @param value    the value to write to the specified location.
      * @param sizeLog2 the size of the value to write. See {@link Sizes}.
      * @throws MemoryAccessException if an error occurred accessing the memory a the specified location.
      */
-    void store(final int address, final long value, final int sizeLog2) throws MemoryAccessException;
+    void store(final long address, final long value, final int sizeLog2) throws MemoryAccessException;
 }
